create a dynamic url loading a certain component and get data
I created a component that would be loaded whenever the url is inside
localhost/products/<some id>

<some id> is the product's id and it will be grabbed by the component
and used inside it.
To achieve that, the dynamic component should be named using square
brackets with an term inside, like [someid].vue

Inside the component, I can use the destructuring feature to get the
term in the url, inside the useRoute().params, the <someid> and use
it inside the component.
The destructuring must have he same name as the component.