use nuxt-link premade component to navigate
the nuxt-link premade component works just like the router-link in the
vue-router. It creates an anchor tag but prevent the http request which
is important to have a more efficient and faster navigation and add
the classes that can be used to style elements differently, depending
on what is the current component.