Add specific layot to specific components
To add code that will be loaded in every component, add a root folder
called layout. Then, add a file called default.vue. The default
component will be loadd in all other pages. Inside the default component
inside the layout folder, add the <slot /> component that will load
the page/component automatically.

To add a specific layout to a certain component, add a new file inside
the layout folder and give it a name, any name.
Then, ad the layout with the <slot /> premade component. Then, in every
page/component you want to put that layout, add the function
definePageMeta() that receives an object. Inside that object, add the
value pair:
laout: '<name of the layout file>'