fetch data for each product using useFetch(link, { key: id })
I'm fetching data for each product. The useFetch() function avoids
making a http request everytime the [id] component is used so it
doesn't work as expected. All product pages show the information for the
same product, made in the first http request. To make the useFetch()
perform one request for each product id, I add an object as the second
argument with a unique id. It differentiate each http request.