crate an error component
Everytime the user tries to access a page that doesn't exist, Nuxt
shows a default error component and send an error props to the error
page.
By creating an "error.vue" component in the root folder, it replaces
the default error component and I can use the error props to display
some information.