Add the app object with metatags like description, title and links
 can add metatags, external links and other stuff in the nuxt.config.ts
 file this way.
inside the defineNuxtConfig export, insert the app value which acceps
an object
this object accept the head value  which also expects an object
inside the head object, I can add the title, in a string and metatags.
the meta expect an array of objects. each object is a metatag that gets
the name and the content
also, inside the head object, I can add external links to stylesheets
and etc with the link value which expects an array of objects
each object in the array is a different link with a rel and the href